backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
openai==0.28.1
python-decouple==3.8
drf-spectacular==0.27.0
python-dotenv==1.0.1
google-generativeai==0.3.2 